Changed to processIntersection and isolated it
